Making subclass of EvergreenService handle state transition in run()
Fix the bug of both EvergreenService and GenericExternalService trying
to handle state transition and cause conflict.
GenericExternalService implement run() in non-blocking way, causes
parent class calling run() returns early and set state to Finished.

Modified to make subclass handle state transition
